refactor tests with PO and facade
