+++ tests/article.spec.js
@@ -1,17 +1,14 @@
 import { test, expect } from '@playwright/test';
-import { faker } from '@faker-js/faker';
-import { EditorArticlePage } from '../src/pages/editorArticlePage.js';
-import { MainPage } from '../src/pages/mainPage.js';
-import { RegisterPage } from '../src/pages/registerPage.js';
-import { ArticlePage } from '../src/pages/articlePage.js';
-import { YourfeedPage } from '../src/pages/yourfeedPage.js';
+import { ArticlePage, EditorArticlePage, MainPage, RegisterPage, YourfeedPage } from '../src/pages/index';
+import { ArticleBuilder, UserBuilder, CommentBuilder } from '../src/helpers/builder/index';
 
 const  URL_UI = 'https://realworld.qa.guru/';
-const user = {
-    username:  faker.person.firstName(),
-    email:  faker.internet.email(),
-    password: faker.internet.password({ length: 10 }),
-};
+
+const userBuilder = new UserBuilder()
+    .addEmail()
+    .addUsername()
+    .addPassword()
+    .generate();
 
 test.describe('Тесты Статьи', () => {
 
@@ -22,7 +19,7 @@ test.describe('Тесты Статьи', () => {
         await mainPage.open(URL_UI);
         await mainPage.gotoRegister();
 
-        await registerPage.register(user.username, user.email, user.password);
+        await registerPage.register(userBuilder.username, userBuilder.email, userBuilder.password);
     });
 
  
@@ -32,19 +29,19 @@ test.describe('Тесты Статьи', () => {
         const editorArticlePage = new EditorArticlePage(page);
         const yourfeedPage = new YourfeedPage(page);
 
-        const article = {
-            title:  faker.book.title() + '_ivandurian',
-            description:  faker.lorem.paragraph(1),
-            body: faker.lorem.paragraph(5),
-            tags: 'реклама'
-        };
+        const articleBuilder = new ArticleBuilder()
+            .addTitle()
+            .addDescription()
+            .addBody()
+            .addTags()
+            .generate();
 
         await yourfeedPage.gotoNewArticle();
-        await editorArticlePage.fillForm(article);
+        await editorArticlePage.fillForm(articleBuilder);
         await editorArticlePage.publish();
 
         await expect(editorArticlePage.header).toBeVisible();
-        await expect(editorArticlePage.header).toContainText(article.title);
+        await expect(editorArticlePage.header).toContainText(articleBuilder.title);
     })
 
     test('Пользователь может оставить комментарий к первой статье в Global Feed', async ({ 
@@ -53,15 +50,17 @@ test.describe('Тесты Статьи', () => {
         const articlePage = new ArticlePage(page);
         const yourfeedPage = new YourfeedPage(page);
 
-        const comment = faker.lorem.sentence()
+        const commentBuilder = new CommentBuilder()
+            .addText()
+            .generate();
 
         await yourfeedPage.openGlobalArticle();
-        await articlePage.addComment(comment);
+        await articlePage.addComment(commentBuilder);
 
         await expect(articlePage.commentText).toBeVisible();
-        await expect(articlePage.commentText).toContainText(comment);
+        await expect(articlePage.commentText).toContainText(commentBuilder);
         await expect(articlePage.commentAuthor).toBeVisible();
-        await expect(articlePage.commentAuthor).toContainText(user.username);
+        await expect(articlePage.commentAuthor).toContainText(userBuilder.username);
     })
 
     test('Пользователь может оставить комментарий к своей статье', async ({ 
@@ -72,43 +71,32 @@ test.describe('Тесты Статьи', () => {
         const editorArticlePage = new EditorArticlePage(page);
         const yourfeedPage = new YourfeedPage(page);
 
-        const article = {
-            title:  faker.book.title() + '_ivandurian',
-            description:  faker.lorem.paragraph(1),
-            body: faker.lorem.paragraph(5),
-            tags: 'реклама'
-        };
+        const articleBuilder = new ArticleBuilder()
+            .addTitle()
+            .addDescription()
+            .addBody()
+            .addTags()
+            .generate();
 
         await yourfeedPage.gotoNewArticle();
-        await editorArticlePage.fillForm(article);
+        await editorArticlePage.fillForm(articleBuilder);
         await editorArticlePage.publish();
 
         await expect(editorArticlePage.header).toBeVisible();
-        await expect(editorArticlePage.header).toContainText(article.title);
+        await expect(editorArticlePage.header).toContainText(articleBuilder.title);
 
         // добавление комментария
         const articlePage = new ArticlePage(page);
 
-        const comment = faker.lorem.sentence()
+        const commentBuilder = new CommentBuilder()
+            .addText()
+            .generate();
 
-        await articlePage.addComment(comment);
+        await articlePage.addComment(commentBuilder);
 
         await expect(articlePage.commentText).toBeVisible();
-        await expect(articlePage.commentText).toContainText(comment);
+        await expect(articlePage.commentText).toContainText(commentBuilder);
         await expect(articlePage.commentAuthor).toBeVisible();
-        await expect(articlePage.commentAuthor).toContainText(user.username);
-
-        /*
-        const articlePage = new ArticlePage(page);
-
-        const comment = faker.lorem.sentence()
-
-        await articlePage.open();
-        await articlePage.addComment(comment);
-
-        await expect(articlePage.card.locator('p.card-text')).toBeVisible();
-        await expect(articlePage.card.locator('p.card-text')).toContainText(comment);
-        await expect(articlePage.card.locator('a.comment-author').last()).toBeVisible();
-        await expect(articlePage.card.locator('a.comment-author').last()).toContainText(user.username);
-    */})
+        await expect(articlePage.commentAuthor).toContainText(userBuilder.username); 
+    })
 });
